requirements.txt
asgiref
billiard
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
Django
django-cors-headers
django-environ
django-filter
djangorestframework
djangorestframework_simplejwt
drf-yasg
idna
inflection
joblib
kombu
marshmallow
numpy
packaging
pillow
prompt_toolkit
psycopg2-binary
PyJWT
python-dateutil
python-dotenv
pytz
PyYAML
redis
requests
scikit-learn
scipy
six
sqlparse
sslcommerz-sdk
stripe
threadpoolctl
typing_extensions
tzdata
uritemplate
urllib3
vine
wcwidth
whitenoise